Backend/requirements.txt
Flask
opencv-python
Flask-CORS
google-generativeai
numpy
requests
mediapipe
numpy
csv